cleaned up unused code
