+++ grails-app/i18n/messages.properties
@@ -278,18 +278,6 @@ expression.test.type.label=Type
 expression.test.button.label=See values
 expression.invalid.exception=Exception was {0}
 
-# calculation info
-info.calculation.average=Average value
-info.calculation.scores=Scores
-info.calculation.trend=Trend
-info.location.na=N/A
-info.normalizeddataelement.missingvalue=Some values are missing
-info.normalizeddataelement.range=Range
-info.normalizeddataelement.expression=Expression
-info.normalizeddataelement.data=Data
-info.normalizeddataelement.na=N/A
-info.normalizeddataelement.trend=Trend
-
 # tasks
 task.label=Task
 task.user.label=User
