+++ grails-app/i18n/messages_fr.properties
@@ -278,18 +278,6 @@ expression.test.type.label=Type
 expression.test.button.label=Voir valeurs
 expression.invalid.exception=Exception: {0}
 
-# calculation info
-info.calculation.average=Valeur moyenne
-info.calculation.scores=Scores
-info.calculation.trend=Tendance
-info.location.na=N/A
-info.normalizeddataelement.missingvalue=Certaines valeurs manquent
-info.normalizeddataelement.range=Etendue
-info.normalizeddataelement.expression=Expression
-info.normalizeddataelement.data=Donnée
-info.normalizeddataelement.na=N/A
-info.normalizeddataelement.trend=Tendance
-
 # tasks
 task.label=Task
 task.user.label=User
