+++ grails-app/services/org/chai/kevin/imports/DataImporter.java
@@ -259,5 +259,4 @@ public abstract class DataImporter extends FileImporter {
 		}
 	}
 	
-	
 }
