+++ grails-app/views/entity/survey/_questionList.gsp
@@ -33,7 +33,7 @@
 				</td>
 				<td>${question.code}</td>
 				<td><a title="Question" href="#" class="cluetip"
-					rel="${createLink(controller:question.getType().getTemplate(), action:'getDescription', params:[question: question.id])}"
+					rel="${createLink(controller:'question', action:'getDescription', params:[question: question.id])}"
 					onclick="return false;"><g:stripHtml field="${i18n(field: question.names)}" chars="100"/></a></td>
 				<td>${question.getType()}</td>
 				<td><g:prettyList entities="${question.typeCodeString}" /></td>
